requirements.txt
django==4.2
djangorestframework
django-cors-headers
openai
python-dotenv
